comment_processor/requirements.txt
openai
pika
python-dotenv
